HaTeXExtension.Math

Description

This module provides those TeX commands, that are usually used in math mode. Also it makes LaTeX type an instance of Num and Frac classes and derives Eq instance.

in command

in_ :: LaTeX

from :: LaTeX -> LaTeX -> LaTeX

from and (&#x2208) (Unicode: U+2208) are operators for in_ command (not in because it is haskell keyword)

(∈) :: LaTeX -> LaTeX -> LaTeX

notin command

notin :: LaTeX

notFrom :: LaTeX -> LaTeX -> LaTeX

notFrom and (&#x2209) (Unicode: U+2209) are operators for notin command

(∉) :: LaTeX -> LaTeX -> LaTeX

Parenthesis, brackets and braces

left :: Char -> LaTeX

\left modificator command

right :: Char -> LaTeX

\right modificator command

leftRight :: (Char, Char) -> LaTeX -> LaTeX

combines left and right functions to surround given statement

parenthesis :: LaTeX -> LaTeX

Encloses expression in \left( and \right)

prn :: LaTeX -> LaTeX

Shothand for parenthesis function

brackets :: LaTeX -> LaTeX

Encloses expression in \left[ and \right]

braces :: LaTeX -> LaTeX

Encloses expression in \left{ and \right}

Application operator

(.:) :: LaTeX -> LaTeX -> LaTeX

Application operator - encloses second arg in parenthesis. Example: f.:(x + y) means f \left( x + y \right) in TeX.

⁖ :: LaTeX -> LaTeX -> LaTeX

Unicode synonym for application operator (.:) (Unicode: U+2056)

Compare commands

(==:) :: LaTeX -> LaTeX -> LaTeX

Just equal sign "=" as operator

leq command

leq :: LaTeX

(<=:) :: LaTeX -> LaTeX -> LaTeX

(≤) :: LaTeX -> LaTeX -> LaTeX

(<=:) and (&#x2264) (Unicode: U+2264) are operators for leq command

geq command

geq :: LaTeX

(>=:) :: LaTeX -> LaTeX -> LaTeX

(≥) :: LaTeX -> LaTeX -> LaTeX

(>=:) and (&#x2265) (Unicode: U+2265) are operators for geq command

Subscript and superscript

⁀ :: LaTeX -> LaTeX -> LaTeX

Just a synonym for (^:) superscript operator (Unicode: U+2040)

‿ :: LaTeX -> LaTeX -> LaTeX

Just a synonym for (!:) subscript operator (Unicode: U+203F)

Other symbols and operators

frac :: LaTeX -> LaTeX -> LaTeX

Just \frac comand

forall :: LaTeX

Just symbol forall

(⎕) :: LaTeX -> LaTeX -> LaTeX

Math medium space command \; that is 5/18 of quad (Unicode: U+2395)