HaTeXExtension.Meta

Description

This module offers templates, that HaTeXExtension uses to simplify declarations of additional commands

Documentation

defTeXCommand

Arguments

:: String
function name

-> String
TeX command name

-> [String]
operator names

-> Q [Dec]
top-level declaration

Template defTeXCommand "foo" "bar" ["+#&","qux"] declares

value foo, that represents TeX command bar without arguments:

 foo :: LaTeX
 foo = TeXComm "bar" []

operator:

 (+#&) :: LaTeX -> LaTeX -> LaTeX
 a +#& b = a <> foo <> b

function:

 qux :: LaTeX -> LaTeX -> LaTeX
 a `qux` b = a <> foo <> b

Template can take any count of such operators-synonims. If first argument is "", then function name will be the same as second argument - command name. This is top-level declaration, so it can be used without special splicing $(...) syntax.

defFusedOperators

Arguments

:: [String]
strings with operator names

-> Q [Dec]
top-level declaration

Template defFusedOperators ["< ... <"] declares operators, that are combinations of several others:

 <...< :: LaTeX -> LaTeX -> LaTeX
 a <...< b = a < "" ... "" < b

This is top-level declaration, so it can be used without special splicing $(...) syntax.